add create post by blog by blogger & tests

+++ src/modules/blogs/use-case/delete.blog.use-case.ts
@@ -1,16 +1,23 @@
+import { ForbiddenException, NotFoundException } from '@nestjs/common';
 import { CommandHandler, ICommandHandler } from '@nestjs/cqrs';
 
 import { BlogsRepo } from '../repositories/blogs.repo';
 
 export class DeleteBlogCommand {
-  constructor(public blogId: string) {}
+  constructor(public userId: string, public blogId: string) {}
 }
 
 @CommandHandler(DeleteBlogCommand)
 export class DeleteBlogUseCase implements ICommandHandler<DeleteBlogCommand> {
   constructor(private blogsRepo: BlogsRepo) {}
 
-  async execute({ blogId }: DeleteBlogCommand) {
+  async execute({ userId, blogId }: DeleteBlogCommand) {
+    const blog = await this.blogsRepo.checkBlogById(blogId);
+
+    if (!blog) throw new NotFoundException();
+
+    if (userId !== blog.user_id) throw new ForbiddenException();
+
     return this.blogsRepo.deleteBlog(blogId);
   }
 }
